feat: Added a new SimpleChat component, replaced the Assistant component, and improved API key management and MCP tool integration. Added uvicorn dependencies to pyproject.toml and uv.lock.
