+++ pyproject.toml
@@ -8,6 +8,7 @@ dependencies = [
     "mcp>=1.7.1",
     "pydantic>=2.11.3",
     "pywin32==310; sys_platform == 'win32'",
+    "uvicorn>=0.34.2",
 ]
 
 [dependency-groups]
